moving compute to its own folder.. updates to appcontext.ts, shrinkwrap.yaml., launch.json
