+++ .vscode/launch.json
@@ -6,29 +6,14 @@
   "configurations": [
     {
       "type": "node",
-      "request": "attach",
-      "name": "Attach",
-      "port": 9229,
-      "smartStep": true,
-      "skipFiles": [
-        "<node_internals>/**/*.js"
-      ],
-    },
-    {
-      "type": "node",
       "protocol": "inspector",
       "request": "launch",
-      "name": "run test",
-      "program": "${workspaceRoot}/adl/node_modules/mocha/bin/_mocha",
-      "args": [
-        "dist/test",
-        "--timeout=2000000"
-      ],
+      "name": "list",
+      "program": "f:/adl/cli/dist/index.js",
+      "args": ["list"],
       "smartStep": true,
-      "skipFiles": [
-        "<node_internals>/**/*.js"
-      ],
-      "cwd": "${workspaceRoot}/adl"
-    },
+      "skipFiles": ["<node_internals>/**/*.js"],
+      "cwd": "${workspaceRoot}"
+    }
   ]
 }
