+++ compute_rp/20180601/module.ts
@@ -21,7 +21,7 @@ import * as core from "../normalized/module";
 // note: It does not require any special conversion logic so wre are composing it with  AutoVersioned<T>
 // Auto converter can convert between versions based on field names.
 export type vmscaleset_resource_20180601 = adltypes.ApiType<
-  "vmscaleset",
+  "vmscalesetfoo",
   "virtual machine scale set",
   core.VirtualMachineNormalized, // [sanjaiga-todo] update this VMScaleset normalized type
   VirtualMachineScaleSet20180601
