+++ compute_rp/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "compute_rp",
+  "version": "1.0.0",
+  "description": "",
+  "main": "api-service.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc -p .",
+    "run": "echo 'run command is not configured'"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@azure-tools/adl.types": "~1.0.0",
+    "@azure-tools/adl.runtime": "~1.0.0",
+    "@azure-tools/arm.adl": "~1.0.0",
+    "@microsoft/ts-command-line": "4.3.10"
+  },
+  "dependencies": {
+    "@azure-tools/adl.types": "~1.0.0",
+    "@azure-tools/adl.runtime": "~1.0.0",
+    "@azure-tools/arm.adl": "~1.0.0",
+    "@microsoft/ts-command-line": "4.3.10"
+  }
+}
